requirements_dev.txt
flake8
tox
coverage
pytest
pytest-runner
jupyter-cadquery==2.0.0
black
pylint


# installs local examples package
-e ./examples
